@@ -1,8 +1,12 @@
 module github.com/foxcpp/go-mockdns
 
-go 1.13
+go 1.21
+
+require github.com/miekg/dns v1.1.57
 
 require (
-	github.com/miekg/dns v1.1.57
+	golang.org/x/mod v0.14.0 // indirect
+	golang.org/x/net v0.18.0 // indirect
+	golang.org/x/sys v0.14.0 // indirect
 	golang.org/x/tools v0.15.0 // indirect
 )
