@@ -6,16 +6,33 @@ import (
 	"log"
 	"net"
 	"os"
+	"strconv"
 	"strings"
+	"sync"
 	"time"
 
 	"github.com/miekg/dns"
+)
+
+// These constants are re-exported from github.com/miekg/dns for convenience,
+// so that users of this package don't need to import github.com/miekg/dns just
+// for these.
+const (
+	RRTypeA     = dns.TypeA
+	RRTypeAAAA  = dns.TypeAAAA
+	RRTypeCNAME = dns.TypeCNAME
+	RRTypeMX    = dns.TypeMX
+	RRTypeNS    = dns.TypeNS
+	RRTypePTR   = dns.TypePTR
+	RRTypeSRV   = dns.TypeSRV
+	RRTypeTXT   = dns.TypeTXT
 )
 
 // Server is the wrapper that binds Resolver to the DNS server implementation
 // from github.com/miekg/dns. This allows it to be used as a replacement
 // resolver for testing code that doesn't support DNS callbacks. See PatchNet.
 type Server struct {
+	mu      sync.RWMutex
 	r       Resolver
 	stopped bool
 	tcpServ dns.Server
@@ -34,6 +51,9 @@ func NewServer(zones map[string]Zone, authoritative bool) (*Server, error) {
 }
 
 func NewServerWithLogger(zones map[string]Zone, l Logger, authoritative bool) (*Server, error) {
+	if zones == nil {
+		zones = make(map[string]Zone)
+	}
 	s := &Server{
 		r: Resolver{
 			Zones: zones,
@@ -77,27 +97,26 @@ func (s *Server) writeErr(w dns.ResponseWriter, reply *dns.Msg, err error) {
 	//reply.Answer = nil
 	reply.Extra = nil
 
-	if dnsErr, ok := err.(*net.DNSError); ok {
-		if isNotFound(dnsErr) {
-			reply.Rcode = dns.RcodeNameError
-			reply.RecursionAvailable = true
-			reply.Ns = []dns.RR{
-				&dns.SOA{
-					Hdr: dns.RR_Header{
-						Name:   dnsErr.Name,
-						Rrtype: dns.TypeSOA,
-						Class:  dns.ClassINET,
-						Ttl:    9999,
-					},
-					Ns:      "localhost.",
-					Mbox:    "hostmaster.localhost.",
-					Serial:  1,
-					Refresh: 900,
-					Retry:   900,
-					Expire:  1800,
-					Minttl:  60,
+	var dnsErr *net.DNSError
+	if errors.As(err, &dnsErr) && dnsErr.IsNotFound {
+		reply.Rcode = dns.RcodeNameError
+		reply.RecursionAvailable = true
+		reply.Ns = []dns.RR{
+			&dns.SOA{
+				Hdr: dns.RR_Header{
+					Name:   dnsErr.Name,
+					Rrtype: dns.TypeSOA,
+					Class:  dns.ClassINET,
+					Ttl:    9999,
 				},
-			}
+				Ns:      "localhost.",
+				Mbox:    "hostmaster.localhost.",
+				Serial:  1,
+				Refresh: 900,
+				Retry:   900,
+				Expire:  1800,
+				Minttl:  60,
+			},
 		}
 	} else {
 		s.Log.Printf("lookup error: %v", err)
@@ -168,6 +187,9 @@ func (s *Server) ServeDNS(w dns.ResponseWriter, m *dns.Msg) {
 		}
 		return
 	}
+
+	s.mu.RLock()
+	defer s.mu.RUnlock()
 
 	switch q.Qtype {
 	case dns.TypeA:
@@ -278,6 +300,7 @@ func (s *Server) ServeDNS(w dns.ResponseWriter, m *dns.Msg) {
 				},
 				Priority: srv.Priority,
 				Port:     srv.Port,
+				Weight:   srv.Weight,
 				Target:   srv.Target,
 			})
 		}
@@ -424,4 +447,143 @@ func (s *Server) Close() error {
 	s.udpServ.Shutdown()
 	s.stopped = true
 	return nil
+}
+
+// Reset removes all zones from the server. The function is thread-safe.
+func (s *Server) Reset() {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+	s.r.Zones = make(map[string]Zone)
+}
+
+// AppendRR appends a resource record to the zone for the given name. If the
+// zone does not exist, it is created. For RR types that only support a single
+// value (CNAME), the existing value is replaced. The function is thread-safe.
+func (s *Server) AppendRR(name string, rrType uint16, rrData string) error {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+
+	if !strings.HasSuffix(name, ".") {
+		name = name + "."
+	}
+	zone := s.r.Zones[name]
+	switch rrType {
+	case dns.TypeA:
+		zone.A = append(zone.A, rrData)
+
+	case dns.TypeAAAA:
+		zone.AAAA = append(zone.AAAA, rrData)
+
+	case dns.TypeTXT:
+		zone.TXT = append(zone.TXT, rrData)
+
+	case dns.TypePTR:
+		zone.PTR = append(zone.PTR, rrData)
+
+	case dns.TypeCNAME:
+		zone.CNAME = rrData
+
+	case dns.TypeMX:
+		parts := strings.SplitN(rrData, " ", 2)
+		if len(parts) != 2 {
+			return errors.New("invalid MX rrData format")
+		}
+		pref, err := strconv.Atoi(parts[0])
+		if err != nil {
+			return errors.New("invalid MX priority")
+		}
+		mx := net.MX{
+			Host: parts[1],
+			Pref: uint16(pref),
+		}
+		zone.MX = append(zone.MX, mx)
+
+	case dns.TypeNS:
+		zone.NS = append(zone.NS, net.NS{Host: rrData})
+
+	case dns.TypeSRV:
+		parts := strings.SplitN(rrData, " ", 4)
+		if len(parts) != 4 {
+			return errors.New("invalid SRV rrData format")
+		}
+		priority, err := strconv.Atoi(parts[0])
+		if err != nil {
+			return errors.New("invalid SRV priority")
+		}
+		weight, err := strconv.Atoi(parts[1])
+		if err != nil {
+			return errors.New("invalid SRV priority")
+		}
+		port, err := strconv.Atoi(parts[2])
+		if err != nil {
+			return errors.New("invalid SRV port")
+		}
+		srv := net.SRV{
+			Priority: uint16(priority),
+			Weight:   uint16(weight),
+			Port:     uint16(port),
+			Target:   parts[3],
+		}
+		zone.SRV = append(zone.SRV, srv)
+
+	default:
+		return errors.New("RR type not supported")
+	}
+	s.r.Zones[name] = zone
+	return nil
+}
+
+// RemoveRR removes all records of the given type from the zone for the given
+// name. If the zone becomes empty, it is removed. The function is thread-safe.
+func (s *Server) RemoveRR(name string, rrType uint16) {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+
+	if !strings.HasSuffix(name, ".") {
+		name = name + "."
+	}
+	zone := s.r.Zones[name]
+	switch rrType {
+	case dns.TypeA:
+		zone.A = nil
+
+	case dns.TypeAAAA:
+		zone.AAAA = nil
+
+	case dns.TypeTXT:
+		zone.TXT = nil
+
+	case dns.TypePTR:
+		zone.PTR = nil
+
+	case dns.TypeCNAME:
+		zone.CNAME = ""
+
+	case dns.TypeMX:
+		zone.MX = nil
+
+	case dns.TypeNS:
+		zone.NS = nil
+
+	case dns.TypeSRV:
+		zone.SRV = nil
+
+	}
+	if isZoneEmpty(zone) {
+		delete(s.r.Zones, name)
+		return
+	}
+	s.r.Zones[name] = zone
+}
+
+func isZoneEmpty(zone Zone) bool {
+	return len(zone.A) == 0 &&
+		len(zone.AAAA) == 0 &&
+		len(zone.TXT) == 0 &&
+		len(zone.PTR) == 0 &&
+		zone.CNAME == "" &&
+		len(zone.MX) == 0 &&
+		len(zone.NS) == 0 &&
+		len(zone.SRV) == 0 &&
+		len(zone.Misc) == 0
 }
